# crm/docs/openapi-mvp.yaml
openapi: 3.1.0
info:
  title: CRM Automotivo MVP API
  version: 0.1.0
  description: Contrato inicial da API Fastify do CRM automotivo MVP.
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
security:
  - bearerAuth: []
...
paths:
...
  /inventory/{id}:
    get: { tags: [Inventory], summary: Detalha estoque., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
    patch: { tags: [Inventory], summary: Atualiza estoque/veiculo., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
  /inventory/{id}/costs:
    post: { tags: [Inventory], summary: Adiciona custo ao veiculo., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "201": { $ref: "#/components/responses/Created" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Ok:
      description: Resposta bem-sucedida.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
    Created:
      description: Recurso criado.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
...